Confirmation dialog on delete

+++ angular-front/src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,8 @@ import { MatDialog } from '@angular/material/dialog';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import {ActivatedRoute, Router} from "@angular/router";
 import {MatSnackBar} from "@angular/material/snack-bar";
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
+
 
 @Component({
   selector: 'app-courses',
@@ -52,16 +54,24 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe(
-      () => {
-        this.fecthData();
-        this.snackBar.open('Curso deletado', 'X', {
-          duration: 5000,
-          verticalPosition: 'top',
-          horizontalPosition: 'center'
-        });
-      }, () => { this.openError("Erro ao remover curso")}
-    );
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Remover curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if(result) {
+        this.coursesService.remove(course._id).subscribe(
+          () => {
+            this.fecthData();
+            this.snackBar.open('Curso deletado', 'X', {
+              duration: 5000,
+              verticalPosition: 'top',
+              horizontalPosition: 'center'
+            });
+          }, () => { this.openError("Erro ao remover curso")}
+        );
+      }
+    });
   }
 
 }
